requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.2.0
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
email-validator==1.1.1
expiringdict==1.2.1
fastapi==0.61.1
filelock==3.0.12
Flask==1.1.2
Flask-Dance==3.0.0
flask-restplus==0.13.0
Flask-Testing==0.8.0
google==3.0.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httplib2==0.18.1
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
oauthlib==3.1.0
orjson==3.3.1
phonenumbers==8.12.9
promise==2.3
protobuf==3.13.0
py3-validate-email==0.2.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.6.1
PyJWT==1.7.1
pyrsistent==0.17.3
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
Rx==1.6.1
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SQLAlchemy==1.3.19
starlette==0.13.6
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.25.10
URLObject==2.4.3
uvicorn==0.11.8
uvloop==0.14.0
websockets==8.1
Werkzeug==0.16.0
zipp==3.1.0